+++ FrameLib_Framework/FrameLib_Memory.h
@@ -314,7 +314,7 @@ public:
             
             // Resize
             
-            void resize(bool tagged, size_t size)   { mStorage->resize(tagged, size); }
+            void resize(bool tagged, unsigned long size)   { mStorage->resize(tagged, size); }
             
         private:
             
@@ -337,7 +337,7 @@ public:
         
         // Resize the storage
         
-        void resize(bool tagged, size_t size);
+        void resize(bool tagged, unsigned long size);
 
         // Constructor / Destructor
